% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_convert_linux.R
\name{coord_convert_linux}
\alias{coord_convert_linux}
\title{Convert coordinates found in GPS exif tags}
\usage{
coord_convert_linux(in.path, crs.in = "+init=epsg:4326",
  crs.out = "+init=epsg:27700", csv = TRUE)
}
\arguments{
\item{in.path}{The full path of input photos}

\item{crs.in}{The native coordinate system}

\item{crs.out}{The target coordinate system}

\item{csv}{Write a .csv file with converted coordinates?}

\item{out.path}{The full path for output photos}
}
\description{
Convert coordinates found in GPS exif tags
}
\author{
James P. Duffy
}
