% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_convert_windows.R
\name{coord_convert_windows}
\alias{coord_convert_windows}
\title{Convert coordinates found in GPS exif tags}
\usage{
coord_convert_windows(in.path, crs.in = "+init=epsg:4326",
  crs.out = "+init=epsg:27700", exif.tool.path, csv = TRUE)
}
\arguments{
\item{in.path}{The full path of input photos}

\item{crs.in}{The native coordinate system}

\item{crs.out}{The target coordinate system}

\item{exif.tool.path}{Full path of exiftool.exe location}

\item{csv}{Write a .csv file with converted coordinates?}
}
\description{
Convert coordinates found in GPS exif tags
}
\author{
James P. Duffy
}
