% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exif_coord_convert.R
\name{exif_coord_convert}
\alias{exif_coord_convert}
\title{Convert coordinates found in GPS exif tags}
\usage{
exif_coord_convert(img_path, crs_in = "+init=epsg:4326",
  crs_out = "+init=epsg:27700", csv = TRUE)
}
\arguments{
\item{img_path}{The full path of input photos}

\item{crs_in}{The native coordinate system}

\item{crs_out}{The target coordinate system}

\item{csv}{Write a .csv file with converted coordinates?}
}
\description{
Convert coordinates found in GPS exif tags
}
\author{
James P. Duffy
}
