% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exif_retag.R
\name{exif_retag}
\alias{exif_retag}
\title{Retag images with GPS info from Ardupilot dataflash logs}
\usage{
exif_retag(time_diff = 0, img_path, log_file, proj_name, csv_out,
  leap_secs = seq(10:50))
}
\arguments{
\item{time_diff}{The difference in time between camera and GPS}

\item{img_path}{The location images to be tagged}

\item{log_file}{The full path of the associated dataflash log}

\item{proj_name}{A name for this project (appended to output .csv)}

\item{csv_out}{Output folder for .csv file}

\item{leap_secs}{Number leapseconds in GPS time since the epoch. default
is 18}
}
\description{
Retag images with GPS info from Ardupilot dataflash logs
}
\author{
James P. Duffy
}
