% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exif_retag_linux.R
\name{exif_retag_linux}
\alias{exif_retag_linux}
\title{Retag photos with GPS info from Ardupilot dataflash logs}
\usage{
exif_retag_linux(timediff = 0, photo.folder.path, log.file.path, proj.name,
  csv.out, leapsecs = seq(10:50))
}
\arguments{
\item{photo.folder.path}{The location photos to be tagged}

\item{log.file.path}{The full path of the associated dataflash log}

\item{proj.name}{A name for this project (appended to output .csv)}

\item{csv.out}{Output folder for .csv file}

\item{leapsecs}{Number leapseconds in GPS time since the epoch. default
is 17}

\item{timedirr}{The difference in time between camera and GPS}
}
\description{
Retag photos with GPS info from Ardupilot dataflash logs
}
\author{
James P. Duffy
}
