% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_attitude.R
\name{extract_attitude}
\alias{extract_attitude}
\title{Extract GPS and attitude information for each image position.}
\usage{
extract_attitude(time_diff = 0, img_path, log_file, proj_name, csv_out,
  leap_secs = seq(10:50))
}
\arguments{
\item{time_diff}{The difference in time between camera and GPS}

\item{img_path}{The location image to be tagged}

\item{log_file}{The full path of the associated dataflash log}

\item{proj_name}{A name for this project (appended to output .csv)}

\item{csv_out}{Output folder for .csv file}

\item{leap_secs}{Number leapseconds in GPS time since the epoch. default
is 18}
}
\description{
Extract GPS and attitude information for each image position.
}
\author{
James P. Duffy
}
