% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gps_attitude_windows.R
\name{extract_gps_attitude_windows}
\alias{extract_gps_attitude_windows}
\title{Extract GPS and attitude information for each photo position.}
\usage{
extract_gps_attitude_windows(timediff = 0, photo.folder.path, log.file.path,
  proj.name, csv.out, leapsecs = seq(10:50), exif.tool.path)
}
\arguments{
\item{timediff}{The difference in time between camera and GPS}

\item{photo.folder.path}{The location photos to be tagged}

\item{log.file.path}{The full path of the associated dataflash log}

\item{proj.name}{A name for this project (appended to output .csv)}

\item{csv.out}{Output folder for .csv file}

\item{leapsecs}{Number leapseconds in GPS time since the epoch. default
is 18}

\item{exif.tool.path}{Full path of exiftool.exe location}
}
\description{
Extract GPS and attitude information for each photo position.
}
\author{
James P. Duffy
}
