% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_time_date.R
\name{format_time_date}
\alias{format_time_date}
\title{Format time and date}
\usage{
format_time_date(weeks, mS, epoch_date, leap_secs)
}
\description{
Format time and date
}
\author{
James P. Duffy
}
\keyword{internal}
