% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synthesiser.R
\name{synthesise}
\alias{synthesise}
\title{Synthesise dataflash logs}
\usage{
synthesise(in.folder, out.folder, recursive = FALSE, leapsecs = seq(10:50),
  ...)
}
\arguments{
\item{in.folder}{The location of log files}

\item{out.folder}{The location for output file}

\item{recursive}{Recusively search folders?}

\item{leapsecs}{Number leapseconds in GPS time since the epoch. default
is 17}
}
\description{
Synthesise dataflash logs
}
\author{
James P. Duffy
}

