% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesise.R
\name{synthesise}
\alias{synthesise}
\title{Synthesise dataflash logs}
\usage{
synthesise(log_path, out_path, out_name = "synthesised_Logs",
  recursive = FALSE, leap_secs = seq(10:50), ...)
}
\arguments{
\item{log_path}{The location of log files}

\item{out_path}{The location for output file}

\item{out_name}{The name of the output text file. Default is 'synthesised_Logs'}

\item{recursive}{Recursively search folders?}

\item{leap_secs}{Number leapseconds in GPS time since the epoch. default
is 18}
}
\description{
Synthesise dataflash logs
}
\author{
James P. Duffy
}
