% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesise.R
\name{synthesise}
\alias{synthesise}
\title{Synthesise dataflash logs}
\usage{
synthesise(in.folder, out.folder, out.name = "Synthesised_Logs",
  recursive = FALSE, leapsecs = seq(10:50), ...)
}
\arguments{
\item{in.folder}{The location of log files}

\item{out.folder}{The location for output file}

\item{out.name}{The name of the output text file. Default is 'Synthesised_Logs'}

\item{recursive}{Recursively search folders?}

\item{leapsecs}{Number leapseconds in GPS time since the epoch. default
is 18}
}
\description{
Synthesise dataflash logs
}
\author{
James P. Duffy
}
